+++ CHANGELOG.md
@@ -11,6 +11,7 @@ n/a
 * Changed navigation bar color in `MainActivity` to match the color of the NavigationBarView.
 * Fixed a bug where rotating the app while the `DetailDialog` is open crashed the app and permanently deleted all app data.
 * Fixed a bug where the color of end icons within the `DetailDialog` was not correct.
+* Added option to change the number of the recently edited accounts on home page.
 
 <br/>